requirements.txt
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
fpdf2==2.7.6
